// frontend/src/Models/Customer.ts
class Customer {
	name: string = '';
	customerType: number = 1;
	description: string = '';
	email: string = '';
	phone: string = '';
	address: string = '';
	city: string = '';
	town: string = '';
	country: string = '';
	zipCode: string = '';
	taxId: string = '';
	taxDepartment: string = '';
}


const CustomerBodyModel = new Customer();


export default CustomerBodyModel;